front: extract buildOpSearchQuery from osrdToNge
