+++ front/src/applications/operationalStudies/components/MacroEditor/osrdToNge.ts
@@ -1,12 +1,8 @@
 import { compact } from 'lodash';
 
 import { osrdEditoastApi } from 'common/api/osrdEditoastApi';
-import type {
-  SearchResultItemOperationalPoint,
-  SearchPayload,
-  SearchQuery,
-  TrainScheduleResult,
-} from 'common/api/osrdEditoastApi';
+import type { SearchResultItemOperationalPoint, SearchPayload } from 'common/api/osrdEditoastApi';
+import buildOpSearchQuery from 'modules/operationalPoint/helpers/buildOpSearchQuery';
 import type { AppDispatch } from 'store';
 
 import nodeStore from './nodeStore';
@@ -126,55 +122,6 @@ const DEFAULT_TIME_LOCK: TimeLockDto = {
 };
 
 /**
- * Build a search query to fetch all operational points from their UICs,
- * trigrams and IDs.
- */
-const buildOpQuery = (
-  infraId: number,
-  trainSchedules: TrainScheduleResult[]
-): SearchPayload | null => {
-  const pathItems = trainSchedules.map((schedule) => schedule.path).flat();
-  const pathItemQueries = [];
-  const pathItemSet = new Set<string>();
-  for (const item of pathItems) {
-    let query: SearchQuery;
-    if ('uic' in item) {
-      query = ['=', ['uic'], item.uic];
-      if (item.secondary_code) {
-        query = ['and', query, ['=', ['ch'], item.secondary_code]];
-      }
-    } else if ('trigram' in item) {
-      query = ['=', ['trigram'], item.trigram];
-      if (item.secondary_code) {
-        query = ['and', query, ['=', ['ch'], item.secondary_code]];
-      }
-    } else if ('operational_point' in item) {
-      query = ['=', ['obj_id'], item.operational_point];
-    } else {
-      continue; // track offset, handled by creating an empty node
-    }
-
-    // Avoid including the same query twice in the search payload
-    const key = JSON.stringify(query);
-    if (pathItemSet.has(key)) {
-      continue;
-    }
-
-    pathItemSet.add(key);
-    pathItemQueries.push(query);
-  }
-
-  if (pathItemQueries.length === 0) {
-    return null;
-  }
-
-  return {
-    object: 'operationalpoint',
-    query: ['and', ['=', ['infra_id'], infraId], ['or', ...pathItemQueries]],
-  };
-};
-
-/**
  * Execute the search payload and collect all result pages.
  */
 const executeSearch = async (
@@ -246,7 +193,7 @@ const importTimetable = async (
     (trainSchedule) => trainSchedule.path.length >= 2
   );
 
-  const searchPayload = buildOpQuery(infraId, trainSchedules);
+  const searchPayload = buildOpSearchQuery(infraId, trainSchedules);
   const searchResults = searchPayload ? await executeSearch(searchPayload, dispatch) : [];
 
   const resource = {
